requirement.txt
torch
torchaudio<2.9.0
soundfile
silero-vad
ffmpeg-python
requests
